fix(auth): added check for ECONNREFUSED errors during login
If there is no formatted valid DeployR response do to the server not being
reachable we pass an appropriate error message in the error callback.

The `deployr.js` library should handle errors that do not make it to DeployR
and normalize the error code and message into the same DeployR client response
structure.

closes #3